hijack thread with unexpected prolog
This is a workaround for linux, which produces a special prolog
for JIT functions, with several pages of stack memory allocated for
local variables, due to aggressive inlining.

If hijack happens at the right place within that prolog, it confuses
StackWalkFramesEx to produce incorrect stack pointer to return address
to hijack.

With incorrect return address being used for hijack, HijackThread
will fail to redirect the thread to correct function, and it allows the
thread to continue executing and unwinding the existing callstack.

When the thread finally calls some native function that checks for
safepoint, it would discover that the thread still has the Hijack
state and would call UnhijackThread that overwrites the memory that
incorrect stack pointer points to with the "original" return address.

The problem is: at this point arbitrary number of stack winding &
unwinding has already happened between HijackThread and UnhijackThread
due to hijack adress injected into the stack at the ineffective place,
that overwriting with "original" return address actually corrupts the
stack.